Move parseGeometryPoint to GeoUtils
