+++ geo/src/main/java/org/odk/collect/geo/geopoly/GeoPolyUtils.kt
@@ -1,5 +1,6 @@
 package org.odk.collect.geo.geopoly
 
+import org.odk.collect.geo.GeoUtils.parseGeometryPoint
 import org.odk.collect.maps.MapPoint
 
 object GeoPolyUtils {
@@ -17,25 +18,6 @@ object GeoPolyUtils {
         }
 
         return points
-    }
-
-    @JvmStatic
-    fun parseGeometryPoint(answer: String?): DoubleArray? {
-        if (answer != null && answer.isNotEmpty()) {
-            val sa = answer.trim { it <= ' ' }.split(" ").toTypedArray()
-            return try {
-                doubleArrayOf(
-                    sa[0].toDouble(),
-                    if (sa.size > 1) sa[1].toDouble() else 0.0,
-                    if (sa.size > 2) sa[2].toDouble() else 0.0,
-                    if (sa.size > 3) sa[3].toDouble() else 0.0
-                )
-            } catch (e: Throwable) {
-                null
-            }
-        } else {
-            return null
-        }
     }
 
     /**
