+++ geo/src/test/java/org/odk/collect/geo/GeoUtilsTest.kt
@@ -4,12 +4,14 @@ import android.content.Context
 import android.location.Location
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import junit.framework.TestCase
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.odk.collect.geo.GeoUtils.formatPointsResultString
+import org.odk.collect.geo.GeoUtils.parseGeometryPoint
 import org.odk.collect.maps.MapPoint
 import org.odk.collect.testshared.LocationTestUtils.createLocation
 
@@ -58,5 +60,34 @@ class GeoUtilsTest {
         assertThat(GeoUtils.formatAccuracy(context, 0.01f), equalTo("0.01 m"))
         assertThat(GeoUtils.formatAccuracy(context, 0.10f), equalTo("0.1 m"))
         assertThat(GeoUtils.formatAccuracy(context, 1.1f), equalTo("1.1 m"))
+    }
+
+    @Test
+    fun parseGeometryPointTest() {
+        var gp =
+            parseGeometryPoint("37.45153333333334 -122.15539166666667 0.0 20.0")!!
+        TestCase.assertEquals(37.45153333333334, gp[0])
+        TestCase.assertEquals(-122.15539166666667, gp[1])
+        TestCase.assertEquals(0.0, gp[2])
+        TestCase.assertEquals(20.0, gp[3])
+
+        gp = parseGeometryPoint("37.45153333333334")!!
+        TestCase.assertEquals(37.45153333333334, gp[0])
+        TestCase.assertEquals(0.0, gp[1])
+        TestCase.assertEquals(0.0, gp[2])
+        TestCase.assertEquals(0.0, gp[3])
+
+        gp = parseGeometryPoint(" 37.45153333333334 -122.15539166666667 0.0 ")!!
+        TestCase.assertEquals(37.45153333333334, gp[0])
+        TestCase.assertEquals(-122.15539166666667, gp[1])
+        TestCase.assertEquals(0.0, gp[2])
+        TestCase.assertEquals(0.0, gp[3])
+
+        TestCase.assertEquals(
+            null,
+            parseGeometryPoint("37.45153333333334 -122.15539166666667 0.0 qwerty")
+        )
+        TestCase.assertEquals(null, parseGeometryPoint(""))
+        TestCase.assertEquals(null, parseGeometryPoint(null))
     }
 }
