+++ geo/src/test/java/org/odk/collect/geo/geopoly/GeoPolyUtilsTest.kt
@@ -4,8 +4,8 @@ import junit.framework.TestCase.assertEquals
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
 import org.junit.Test
+import org.odk.collect.geo.GeoUtils.parseGeometryPoint
 import org.odk.collect.geo.geopoly.GeoPolyUtils.parseGeometry
-import org.odk.collect.geo.geopoly.GeoPolyUtils.parseGeometryPoint
 import org.odk.collect.maps.MapPoint
 
 class GeoPolyUtilsTest {
@@ -87,32 +87,6 @@ class GeoPolyUtilsTest {
         )
 
         assertThat(GeoPolyUtils.intersects(trace), equalTo(true))
-    }
-
-    @Test
-    fun parseGeometryPointTest() {
-        var gp =
-            parseGeometryPoint("37.45153333333334 -122.15539166666667 0.0 20.0")!!
-        assertEquals(37.45153333333334, gp[0])
-        assertEquals(-122.15539166666667, gp[1])
-        assertEquals(0.0, gp[2])
-        assertEquals(20.0, gp[3])
-
-        gp = parseGeometryPoint("37.45153333333334")!!
-        assertEquals(37.45153333333334, gp[0])
-        assertEquals(0.0, gp[1])
-        assertEquals(0.0, gp[2])
-        assertEquals(0.0, gp[3])
-
-        gp = parseGeometryPoint(" 37.45153333333334 -122.15539166666667 0.0 ")!!
-        assertEquals(37.45153333333334, gp[0])
-        assertEquals(-122.15539166666667, gp[1])
-        assertEquals(0.0, gp[2])
-        assertEquals(0.0, gp[3])
-
-        assertEquals(null, parseGeometryPoint("37.45153333333334 -122.15539166666667 0.0 qwerty"))
-        assertEquals(null, parseGeometryPoint(""))
-        assertEquals(null, parseGeometryPoint(null))
     }
 
     @Test
